+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use consistent and accessible colour palette across all editors (#173)
+- Allow to select multiple annotations in graph editor as filter (#4)
 
 ### Fixed